Meta data for Disney Touring Plan Park Data

A dataset containing daily details on Disney World Parks

parks_metadata_raw

Format

A data frame with 2079 rows and 181 variables:

date

Park Day (not actual date stamp of the wait time, since some are after midnight), MM/DD/YYY

wdw_ticket_season

Walt Disney World Single Day Price Type, alphanumeric

dayofweek

Day of Week, numeric

dayofyear

Day of Year (0 to 365), numeric

weekofyear

Week of Year (0 to 53), numeric

monthofyear

Month of Year (1 to 12), numeric

year

Year, numeric

season

Seasonal Category (Fall, Winter, Thanksgiving, etc.), alphanumeric

holidaypx

Proximity to Holiday (2-directional) (in days), numeric

holidaym

Holiday Metric (1 to 5 ranking), numeric

holidayn

Name of Holiday, alphanumeric

holiday

Day is a Holiday, Boolean

wdwticketseason

WDW 1-day ticket season, alphanumeric

wdwracen

XXX Event Code, alphanumeric

wdweventn

XXX Event Code, alphanumeric

wdwevent

WDW Event, Boolean

wdwrace

WDW Race, Boolean

wdwseason

WDW Season, alphanumeric

wdwmaxtemp

WDW Max Temperature, numeric

wdwmintemp

WDW Min Temperature, numeric

wdwmeantemp

WDW Average Temperature, numeric

mkeventn

Magic Kingdom Event Code, alphanumeric

mkevent

Magic Kingdom Event, Boolean

epeventn

Epcot Event Code, alphanumeric

epevent

Epcot Event, Boolean

hseventn

Hollywood Studios Event Code, alphanumeric

hsevent

Hollywood Studios Event, Boolean

akeventn

Animal Kingdom Event Code, alphanumeric

akevent

Animal Kingdom Event, Boolean

holidayj

Name of Jewish Holiday, alphanumeric

insession

Percentage of Schools in Session, numeric

insession_enrollment

Percentage of Schools in Session Weighted by Enrollment, numeric

insession_wdw

Percentage of Schools in Session WDW markets only, numeric

insession_dlr

Percentage of Schools in Session DLR markets only, numeric

insession_sqrt_wdw

Percentage of Schools in Session Weighted by Square Root of Distance to WDW, numeric

insession_sqrt_dlr

Percentage of Schools in Session Weighted by Square Root of Distance to DLR, numeric

insession_california

Percentage of Schools in Session California Only, numeric

insession_dc

Percentage of Schools in Session D.C. Only, numeric

insession_central_fl

Percentage of Schools in Session Within Central Florida Only, numeric

insession_drive1_fl

Percentage of Schools in Session Within Driving Distance to Florida Only, numeric

insession_drive2_fl

Percentage of Schools in Session Within Driving Distance to Florida Only, numeric

insession_drive_ca

Percentage of Schools in Session Within Driving Distance to California Only, numeric

insession_florida

Percentage of Schools in Session Florida Only, numeric

insession_mardi_gras

Percentage of Schools in Session Mardi Gras States Only, numeric

insession_midwest

Percentage of Schools in Session Midwest Only, numeric

insession_ny_nj

Percentage of Schools in Session New York and New Jersey Only, numeric

insession_ny_nj_pa

Percentage of Schools in Session New York, New Jersey and Pennsylvania Only, numeric

insession_new_england

Percentage of Schools in Session New England Only, numeric

insession_new_jersey

Percentage of Schools in Session New Jersey Only, numeric

insession_nothwest

Percentage of Schools in Session Northwest Only, numeric

insession_planes

Percentage of Schools in Session Planes Only, numeric

insession_socal

Percentage of Schools in Session Southern California Only, numeric

insession_southwest

Percentage of Schools in Session Southwest Only, numeric

mkemhmorn

Magic Kingdom Extra Magic Hour Morning, Boolean

mkemhmyest

Magic Kingdom Extra Magic Hour Morning Yesterday, Boolean

mkemhmtom

Magic Kingdom Extra Magic Hour Morning Tomorrow, Boolean

mkemheve

Magic Kingdom Extra Magic Hour Evening, Boolean

mkhoursemh

Total Opening Hours including EMH for Magic Kingdom, numeric

mkhoursemhyest

Yesterday\'s Total Opening Hours including EMH for Magic Kingdom, numeric

mkhoursemhtom

Tomorrow\'s Total Opening Hours including EMH for Magic Kingdom, numeric

mkemheyest

Magic Kingdom Extra Magic Hour Evening Yesterday, Boolean

mkemhetom

Magic Kingdom Extra Magic Hour Evening Tomorrow, Boolean

epemhmorn

Epcot Extra Magic Hour Morning, Boolean

epemhmyest

Epcot Extra Magic Hour Morning Yesterday, Boolean

epemhmtom

Epcot Extra Magic Hour Morning Tomorrow, Boolean

epemheve

Epcot Extra Magic Hour Evening, Boolean

epemheyest

Epcot Extra Magic Hour Evening Yesterday, Boolean

epemhetom

Epcot Extra Magic Hour Evening Tomorrow, Boolean

ephoursemh

Total Opening Hours including EMH for Epcot, numeric

ephoursemhyest

Yesterday\'s Total Opening Hours including EMH for Epcot, numeric

ephoursemhtom

Tomorrow\'s Total Opening Hours including EMH for Epcot, numeric

hsemhmorn

Hollywood Studios Extra Magic Hour Morning, Boolean

hsemhmyest

Hollywood Studios Extra Magic Hour Morning Yesterday, Boolean

hsemhmtom

Hollywood Studios Extra Magic Hour Morning Tomorrow, Boolean

hsemheve

Hollywood Studios Extra Magic Hour Evening, Boolean

hsemheyest

Hollywood Studios Extra Magic Hour Evening Yesterday, Boolean

hsemhetom

Hollywood Studios Extra Magic Hour Evening Tomorrow, Boolean

hshoursemh

Total Opening Hours including EMH for Hollywood Studios, numeric

hshoursemhyest

Yesterday\'s Total Opening Hours including EMH for Hollywood Studios, numeric

hshoursemhtom

Tomorrow\'s Total Opening Hours including EMH for Hollywood Studios, numeric

akemhmorn

Animal Kingdom Extra Animal Hour Morning, Boolean

akemhmyest

Animal Kingdom Extra Animal Hour Morning Yesterday, Boolean

akemhmtom

Animal Kingdom Extra Animal Hour Morning Tomorrow, Boolean

akemheve

Animal Kingdom Extra Animal Hour Evening, Boolean

akemheyest

Animal Kingdom Extra Animal Hour Evening Yesterday, Boolean

akemhetom

Animal Kingdom Extra Animal Hour Evening Tomorrow, Boolean

akhoursemh

Total Opening Hours including EMH for Animal Kingdom, numeric

akhoursemhyest

Yesterday\'s Total Opening Hours including EMH for Animal Kingdom, numeric

akhoursemhtom

Tomorrow\'s Total Opening Hours including EMH for Animal Kingdom, numeric

mkopen

Opening Hour for Magic Kingdom, HH:MM

mkclose

Closing Hour for Magic Kingdom, HH:MM

mkhours

Total Opening Hours for Magic Kingdom, numeric

mkemhopen

Magic Kingdom Extra Magic Hour Opening Time, HH:MM

mkemhclose

Magic Kingdom Extra Magic Hour Closing Time, HH:MM

mkopenyest

Yesterday\'s Opening Hour for Magic Kingdom, HH:MM

mkcloseyest

Yesterday\'s Closing Hour for Magic Kingdom, HH:MM

mkhoursyest

Yesterday\'s Total Opening Hours for Magic Kingdom, numeric

mkopentom

Tomorrow\'s Opening Hour for Magic Kingdom, HH:MM

mkclosetom

Tomorrow\'s Closing Hour for Magic Kingdom, HH:MM

mkhourstom

Tomorrow\'s Total Opening Hours for Magic Kingdom, numeric

epopen

Opening Hour for Epcot, HH:MM

epclose

Closing Hour for Epcot, HH:MM

ephours

Total Opening Hours for Epcot, numeric

epemhopen

Epcot Kingdom Extra Magic Hour Opening Time, HH:MM

epemhclose

Epcot Extra Magic Hour Closing Time, HH:MM

epopenyest

Yesterday\'s Opening Hour for Epcot, HH:MM

epcloseyest

Yesterday\'s Closing Hour for Epcot, HH:MM

ephoursyest

Yesterday\'s Total Opening Hours for Epcot, numeric

epopentom

Tomorrow\'s Opening Hour for Epcot, HH:MM

epclosetom

Tomorrow\'s Closing Hour for Epcot, HH:MM

ephourstom

Tomorrow\'s Total Opening Hours for Epcot, numeric

hsopen

Opening Hour for Disney Hollywood Studios, HH:MM

hsclose

Closing Hour for Disney Hollywood Studios, HH:MM

hshours

Total Opening Hours for Disney Hollywood Studios, numeric

hsemhopen

Hollywood Studios Extra Magic Hour Opening Time, HH:MM

hsemhclose

Hollywood Studios Extra Magic Hour Closing Time, HH:MM

hsopenyest

Yesterday\'s Opening Hour for Hollywood Studios, HH:MM

hscloseyest

Yesterday\'s Closing Hour for Hollywood Studios, HH:MM

hshoursyest

Yesterday\'s Total Opening Hours for Hollywood Studios, numeric

hsopentom

Tomorrow\'s Opening Hour for Hollywood Studios, HH:MM

hsclosetom

Tomorrow\'s Closing Hour for Hollywood Studios, HH:MM

hshourstom

Tomorrow\'s Total Opening Hours for Hollywood Studios, numeric

akopen

Opening Hour for Animal Kingdom, HH:MM

akclose

Closing Hour for Animal Kingdom, HH:MM

akhours

Total Opening Hours for Animal Kingdom, numeric

akemhopen

Animal Kingdom Extra Magic Hour Opening Time, HH:MM

akemhclose

Animal Kingdom Extra Magic Hour Closing Time, HH:MM

akopenyest

Yesterday\'s Opening Hour for Animal Kingdom, HH:MM

akcloseyest

Yesterday\'s Closing Hour for Animal Kingdom, HH:MM

akhoursyest

Yesterday\'s Total Opening Hours for Animal Kingdom, numeric

akopentom

Tomorrow\'s Opening Hour for Animal Kingdom, HH:MM

akclosetom

Tomorrow\'s Closing Hour for Animal Kingdom, HH:MM

akhourstom

Tomorrow\'s Total Opening Hours for Animal Kingdom, numeric

weather_wdwhigh

Historical High Temperature, numeric

weather_wdwlow

Historical Low Temperature, numeric

weather_wdwprecip

Historical Precipitation, numeric

capacitylost_mk

Total hourly capacity lost on that park day (due to attraction closures), numeric

capacitylost_ep

Total hourly capacity lost on that park day, numeric

capacitylost_hs

Total hourly capacity lost on that park day, numeric

capacitylost_ak

Total hourly capacity lost on that park day, numeric

capacitylostwgt_mk

Total hourly capacity lost on that park day, weighted by attraction popularity, numeric

capacitylostwgt_ep

Total hourly capacity lost on that park day, weighted by attraction popularity, numeric

capacitylostwgt_hs

Total hourly capacity lost on that park day, weighted by attraction popularity, numeric

capacitylostwgt_ak

Total hourly capacity lost on that park day, weighted by attraction popularity, numeric

mkprdday

Number of Daytime Parades at Magic Kingdom, numeric

mkprddt1

1st Parade Time at Magic Kingdom, HH:MM

mkprddt2

2nd Parade Time at Magic Kingdom, HH:MM

mkprddn

Magic Kingdom Day Time Parade Name, alphanumeric

mkprdngt

Number of Night Time Parades at Magic Kingdom, Boolean

mkprdnt1

1st Night Parade Time at Magic Kingdom, HH:MM

mkprdnt2

2nd Night Parade Time at Magic Kingdom, HH:MM

mkprdnn

Magic Kingdom Night Parade Name, alphanumeric

mkfirewk

Number of Fireworks Shows at Magic Kingdom, numeric

mkfiret1

1st Magic Kingdom Fireworks time, HH:MM

mkfiret2

2nd Magic Kingdom Fireworks time, HH:MM

mkfiren

Evening Fireworks Name at Magic Kingdom, alphanumeric

epfirewk

Number of Fireworks Shows at Epcot, numeric

epfiret1

1st Epcot Fireworks time, HH:MM

epfiret2

2nd Epcot Fireworks time, HH:MM

epfiren

Evening Fireworks Name at Epcot, alphanumeric

hsprdday

Number of Daytime Parades at Hollywood Studios, numeric

hsprddt1

1st Parade Time at Hollywood Studios, HH:MM

hsprddn

Hollywood Studios Day Time Parade Name, alphanumeric

hsfirewk

Number of Fireworks Shows at Hollywood Studios, numeric

hsfiret1

1st Hollywood Studios Fireworks time, HH:MM

hsfiret2

2nd Hollywood Studios Fireworks time, HH:MM

hsfiren

Evening Fireworks Name at Hollywood Studios, alphanumeric

hsshwngt

Number of Night Time Shows at Hollywood Studios, numeric

hsshwnt1

1st Hollywood Studios Night Time Show time, HH:MM

hsshwnt2

2nd Hollywood Studios Night Time Show time, HH:MM

hsshwnn

Night Time Show Name at Hollywood Studios, alphanumeric

hsfirewks

Number of Fireworks Shows at Hollywood Studios, numeric

akprdday

Number of Daytime Parades at Animal Kingdom, numeric

akprddt1

1st Parade Time at Animal Kingdom, HH:MM

akprddt2

2nd Parade Time at Animal Kingdom, HH:MM

akprddn

Animal Kingdom Day Time Parade Name, alphanumeric

akfiren

Fireworks Name at Animal Kingdom, alphanumeric

akshwngt

Number of Night Time Shows at Animal Kingdom, numeric

akshwnt1

1st Animal Kingdom Night Time Show time, HH:MM

akshwnt2

2nd Animal Kingdom Night Time Show time, HH:MM

akshwnn

Night Time Show Name at Animal Kingdom, alphanumeric

Source

"metadata.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, August 2023, https://touringplans.com/walt-disney-world/crowd-calendar#DataSets, Accessed 28 August 2023