Reduced data set for Disney World attraction wait times in 2018

A dataset containing hourly average posted and actual wait times for select Disney World attractions in 2018

touringplans_2018

Format

A data frame with 66,492 rows and 13 variables:

date

Park Day (not actual date stamp of the wait time, since some are after midnight), Date

hour

Hour of recorded wait time, numeric

name

Disney's official attraction name, alphanumeric

avg_sactmin

Average actual Wait Time (in minutes), numeric

avg_spostmin

Average standby Posted Wait Time (in minutes), numeric

average_wait_per_hundred

The average wait per 100 guests in line, numeric

duration

Duration of attraction in minutes, numeric

park

Disney theme park where the attraction is located

land

Section of the park an attraction is located

open

Opening time for park where attraction is located, HH:MM

close

Closing time for park where attraction is located, HH:MM

extra_magic_morning

Extra magic hour morning, Boolean

extra_magic_evening

Extra magic hour evening, Boolean

wdw_ticket_season

Walt Disney World Single Day Price Type, alphanumeric

short_name

Short name, alphanumeric

Source

Disney World Ride Wait Time Datasets, TouringPlans.com, September 2021, https://touringplans.com/walt-disney-world/crowd-calendar#DataSets, Accessed 16 September 2021